Added initial file
Easy to use and implement! (I hope)

Planned updates:

Add comments

Add a patch to a potential bug. (The bug in question is data loss. This can potentially be caused by the engine working on some data for too long and end up missing a refresh or more. This can hopefully be mitigated or removed by having checks for new data become independent from the actual execution code and having it queue packs instead of replacing the current one. Perhaps trying to speed up execution until it's caught up. But I've not seen the bug and haven't proved its existence, soo, yeah.)